+++ dist_node/parse/expression_parser.js
@@ -67,6 +67,9 @@ var __o = require("bennu")["parse"],
 (memberExpression = late((function() {
     return memberExpression;
 })));
+(curryExpression = late((function() {
+    return memberExpression;
+})));
 (arrayLiteral = label("Array Literal", ((arrayElement = expression), (arrayElements = eager(sepBy(punctuator(","),
     arrayElement))), node(between(punctuator("["), punctuator("]"), expected("array element", arrayElements)),
     ast_expression.ArrayExpression.create))));
@@ -108,9 +111,9 @@ var unaryOperatorExpression = label("Unary Operator Expression", bind(either(key
         value = __o["value"];
     return always(ast_expression.UnaryOperatorExpression.create(loc, value));
 }))),
-    binaryOperatorExpression = label("Binary Operator Expression", bind(either(keyword("instanceof"), punctuator("*",
-        "/", "+", "-", "%", "<<", ">>", ">>>", "<", ">", "<=", ">=", "==", "!=", "===", "!==", "&", "^",
-        "|", "||", "&&", "\\>", "|>")), (function(__o) {
+    binaryOperatorExpression = label("Binary Operator Expression", bind(either(keyword("instanceof"), punctuator(".",
+        "*", "/", "+", "-", "%", "<<", ">>", ">>>", "<", ">", "<=", ">=", "==", "!=", "===", "!==", "&",
+        "^", "|", "||", "&&", "|>", "\\>", "\\>>", "<|", "<\\", "<<\\")), (function(__o) {
         var loc = __o["loc"],
             value = __o["value"];
         return always(ast_expression.BinaryOperatorExpression.create(loc, value));
@@ -149,7 +152,8 @@ var unaryOperatorExpression = label("Unary Operator Expression", bind(either(key
     });
 }))));
 (newExpression = label("New Expression", nodea(next(keyword("new"), enumeration(expected("member expression",
-    memberExpression), expected("argument list", args))), ast_expression.NewExpression.create)));
+        memberExpression), expected("argument list", either(args, curryExpression)))), ast_expression.NewExpression
+    .create)));
 var accessorReducer = (function(p, c) {
     return ast_expression.MemberExpression.create(SourceLocation.merge(p.loc, c.loc), p, c.property, c.computed);
 });